move protocol-specific stuff to separate modules
